ramips: Add support for Netgear EX2700
This patch adds support for the Netgear EX2700 and builds an approriate
sysupgrade image.

What's missing is the option to build a factory image flashable via the
router's stock web interface, but this approach is hindered by the fact
that u-boot operforms an additional integrity check, which expects a
uImage header in the last 64 bytes of the "kernel" partition, which
the bootloader expects to be 960k, a size exceeded by the standard
OpenWrt kernel.

Signed-off-by: Joseph C. Lehner <joseph.c.lehner@gmail.com>

git-svn-id: svn://svn.openwrt.org/openwrt/trunk@48698 3c298f89-4303-0410-b956-a3cf2f4a3e73