lantiq: lzma-loader: fix O32 ABI conformance
According to the calling convention of the o32 ABI the
caller function must reserve stack space for $a0-$a3
registers in case the callee needs to save its arguments.

The assembly startup code does not reserve stack space
for these registers thus when the main C function needs
to save its arguments, that will cause a stack overflow.

Fix the assembly code to reserve stack space for the
registers to avoid that.

Compile tested only.

Signed-off-by: Gabor Juhos <juhosg@openwrt.org>

git-svn-id: svn://svn.openwrt.org/openwrt/trunk@49061 3c298f89-4303-0410-b956-a3cf2f4a3e73