ar71xx/image: allow builds of loader without cmdline injecting
On most platforms the cmdline is supplied by the loader to the kernel.
When using an elf loader with cmdline the size is bloated by 512k.
For loader-elf platforms like the cpe510 better use loader-elf but patch
the cmdline into the kernel.

wdr1043: Tested loader still works using sysupgade image.

Signed-off-by: Alexander Couzens <lynxis@fe80.eu>

git-svn-id: svn://svn.openwrt.org/openwrt/trunk@49080 3c298f89-4303-0410-b956-a3cf2f4a3e73