arc770: build kmod-ath9k-htc wpad-mini by default
AXS101 beind a development board lacks built-in wireles inerfaces.
So we have to use external USB dongles to turn the board into
wireless router.

The best USB Wi-Fi dongles to work in AP-mode seem to be based on
ath9k-htc chipset.

And so with that change we add support of mentioned dongles in
default and axs101 builds.

Signed-off-by: Alexey Brodkin <abrodkin@synopsys.com>

SVN-Revision: 49133